Fix some warnings

+++ include/stitcher_params.hpp
@@ -4,8 +4,8 @@
 namespace gpustitch{
 
 struct Stitcher_params{
-	size_t width;
-	size_t height;
+	int width;
+	int height;
 };
 
 }
